Clean Arquitecture, namespaces and responsability distribution update. Just starting, rebuild tests and implement user operations
