+++ ForeignExchange.Domain/Entities/ExchangeRate.cs
