+++ ForeignExchange.Domain/Events/ExchangeRateUpdatedEvent.cs
