+++ ForeignExchange.Domain/Exceptions/CurrencyPairException.cs
