+++ ForeignExchange.Domain/Exceptions/ForexProviderException.cs
