+++ ForeignExchange.Infrastructure/Interfaces/IExchangeRateRepository.cs
