+++ ForeignExchange.Infrastructure/Interfaces/IForexProviderRepository.cs
