+++ ForeignExchange.Infrastructure/Model/ExchangeRateResponse.cs
