+++ ForeignExchange.Services/Interfaces/IAuthService.cs
