+++ ForeignExchange.Services/Interfaces/IMessageService.cs
