+++ ForeignExchange.sln
@@ -7,6 +7,12 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ForeignExchange", "ForeignE
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ForeignExchange.Tests", "ForeignExchange.Tests\ForeignExchange.Tests.csproj", "{19DD1249-4BD2-4CD3-AE42-62B626215FFE}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ForeignExchange.Domain", "ForeignExchange.Domain\ForeignExchange.Domain.csproj", "{406B2AED-C4F6-4482-B4F0-4B8771AEF3FA}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ForeignExchange.Infrastructure", "ForeignExchange.Infrastructure\ForeignExchange.Infrastructure.csproj", "{0C81B63A-DE87-4342-A4DC-8EB2480B1AA5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ForeignExchange.Application", "ForeignExchange.Services\ForeignExchange.Application.csproj", "{029927B5-77D8-487F-849E-C75BC38756FD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +27,18 @@ Global
 		{19DD1249-4BD2-4CD3-AE42-62B626215FFE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{19DD1249-4BD2-4CD3-AE42-62B626215FFE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{19DD1249-4BD2-4CD3-AE42-62B626215FFE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{406B2AED-C4F6-4482-B4F0-4B8771AEF3FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{406B2AED-C4F6-4482-B4F0-4B8771AEF3FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{406B2AED-C4F6-4482-B4F0-4B8771AEF3FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{406B2AED-C4F6-4482-B4F0-4B8771AEF3FA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0C81B63A-DE87-4342-A4DC-8EB2480B1AA5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0C81B63A-DE87-4342-A4DC-8EB2480B1AA5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0C81B63A-DE87-4342-A4DC-8EB2480B1AA5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0C81B63A-DE87-4342-A4DC-8EB2480B1AA5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{029927B5-77D8-487F-849E-C75BC38756FD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{029927B5-77D8-487F-849E-C75BC38756FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{029927B5-77D8-487F-849E-C75BC38756FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{029927B5-77D8-487F-849E-C75BC38756FD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
